Refactor rust code to submodules
