+++ deebot_client/rs/map.pyi
@@ -1,8 +1,5 @@
 from typing import Self
 
-def decompress_7z_base64_data(value: str) -> bytes:
-    """Decompress base64 decoded 7z compressed string."""
-
 class TracePoint:
     """Trace point."""
