+++ deebot_client/rs/util.pyi
@@ -1,0 +1,2 @@
+def decompress_7z_base64_data(value: str) -> bytes:
+    """Decompress base64 decoded 7z compressed string."""
